% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convert.numeric}
\alias{convert.numeric}
\title{Test and convert to numeric}
\usage{
convert.numeric(col)
}
\arguments{
\item{col}{a given vector}
}
\value{
perhaps a numeric vector
}
\description{
Test if a given vector is numeric and if yes, then convert it to
numeric.
}
