% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddplysum.R
\name{ddplygr}
\alias{ddplygr}
\title{Calculate a growth rate}
\usage{
ddplygr(data, variables, naswitch = TRUE)
}
\arguments{
\item{data}{the data.frame}

\item{variables}{the variables to group by}

\item{naswitch}{logical, should we add a NA at the beginning? If
TRUE, we add the growth rate to the original data.frame. If FALSE,
a new data.frame will be returned}
}
\description{
Add a growth rate column for a given column in a data.frame
}

