% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddplysum.R
\name{ddplysum}
\alias{ddplysum}
\title{Sum a column by group}
\usage{
ddplysum(data, variables, tosum, naswitch = FALSE)
}
\arguments{
\item{data}{the data.frame}

\item{variables}{the variables to group by, use .(column1, column2)}

\item{tosum}{the column to be summed}

\item{naswitch}{logical, should NAs be ignored?}
}
\description{
Take a data.frame, separate it by the variables given and calculate
}
