% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fac2char}
\alias{fac2char}
\title{Convert a factor to character}
\usage{
fac2char(factor)
}
\arguments{
\item{factor}{the factor vector}
}
\description{
Given a factor vector, create a character vector from it
}

