% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{fac2num}
\alias{fac2num}
\title{Convert a factor to a number}
\usage{
fac2num(factor)
}
\arguments{
\item{factor}{the factor vector}
}
\description{
Given a factor vector, create a numeric vector from it
}

