% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getInd}
\alias{getInd}
\title{Extract a position in a vector}
\usage{
getInd(x, position)
}
\arguments{
\item{x}{the vector}

\item{position}{the position}
}
\description{
Given a vector, return a vector with all zeros except at 'position'
}

