% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Difference in sets}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\description{
Given two vectors, get the element of the first vector that are NOT
in the second vector
}
