% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.finite.data.frame}
\alias{is.finite.data.frame}
\title{Finite data.frame}
\usage{
\method{is.finite}{data.frame}(df)
}
\arguments{
\item{df}{the data.frame}
}
\description{
Test if a data.frame only contains finite values
}
