% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/los.R
\name{los}
\alias{los}
\title{List loaded objects}
\usage{
los(..., n = 10)
}
\arguments{
\item{...}{arguments passed onto lower level functions, no idea
what exactly}

\item{n}{the first 'n' objects are displayed}
}
\description{
List all objects in enviroment, but prettily
}

