% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergehelp.R
\name{mergehelp}
\alias{mergehelp}
\title{Show corresponding (or not) factor levels}
\usage{
mergehelp(x, y, by.x, by.y = by.x)
}
\arguments{
\item{x}{first dataset}

\item{y}{two dataset}

\item{by.x}{factor in first dataset}

\item{by.y}{fatcor in second dataset}
}
\description{
Given two datasets, show what factor levels coincide, and which can
only be found in one of the two datasets
}

