% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sampleview}
\alias{sampleview}
\title{Display a 'n' random rows of a data.frame}
\usage{
sampleview(df, n = 6)
}
\arguments{
\item{df}{the data.frame}

\item{n}{the number of rows to be displayed}
}
\description{
Given a data.frame, display a random selection of it. To see more
than just the head or tail of the data
}

