% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seqsum.R
\name{seqsum}
\alias{seqsum}
\title{Sequential summing}
\usage{
seqsum(mat, n = 3)
}
\arguments{
\item{mat}{a vector or one-dimensional matrix}

\item{n}{number of elements that are summed up}
}
\description{
Given a vector or a matrix, sum every 'n' element up
}
\examples{
x <- rnorm(15)
seqsum(x, 3)
}

