% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDFRow.R
\name{splitDTRow}
\alias{splitDTRow}
\title{Split a rows of a data.table into several ones, using a character field}
\usage{
splitDTRow(dt, splitby = ",", column)
}
\arguments{
\item{dt}{the data.table to be splitted}

\item{splitby}{a character sequence which will act as the separator
for the character field/column}

\item{column}{the column where the character codes are}
}
\description{
From a character field with comma-separated codes, split the row in
several rows, one for each code in the character field
}
\examples{
dta <- data.table(a = 1:2, b = c("a, b, c", "d, e"), stringsAsFactors = FALSE)
dta2 <- splitDTRow(dta, ", ", "b")
}
