% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripAttr.R
\name{unsetAllAttr}
\alias{unsetAllAttr}
\title{Remove all attributes from all columns of a data.table}
\usage{
unsetAllAttr(dt)
}
\arguments{
\item{dt}{a data.table}
}
\value{
a cleaned data.table
}
\description{
Removes all attributes of all columns as they are often unnecessary. Found on stackoverflow and extendeed.
}
