% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_github_pkgs.R
\name{update_github_pkgs}
\alias{update_github_pkgs}
\title{Function to update packages installed from github}
\usage{
update_github_pkgs()
}
\description{
This function was copied from the link in the references.
I only changed the return type.
}
\seealso{
http://stackoverflow.com/questions/32538052/update-all-packages-from-github
}
