% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{To view a data.frame / data.table in LibreOffice Calc}
\usage{
view(data)
}
\arguments{
\item{data}{the data.frame (or data.table)}
}
\description{
Take a data.frame, save it to a temporary file and open it with the
spreadsheet viewer of the OS.
Requires the package 'openxlsx' to work properly.
}

